Refactor voteAutocomplete not to need a HOC
